+++ js/data.js
@@ -1,0 +1,8 @@
+const COMMENT = ['Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+const NAME = ['Алексей', 'Маша', 'Витя', 'Саша', 'Марина', 'Люба', 'Слава', 'Федор', 'Лиля', 'Семен'];
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MIN_AVATAR = 1;
+const MAX_AVATAR = 6;
+
+export {COMMENT, NAME, MIN_LIKES, MAX_LIKES, MIN_AVATAR, MAX_AVATAR};
